Avoid unusable system state during apt install
Installing ROS 2 on a system which hasn't get upgraded the installed
packages from the versions published at Jammy release to those published
in subsequent updates can result in undesirable behavior of the apt
resolver.

The dependency solution that apt presents seems to include the removal
of critical system packages. If the user doesn't notice the misbehavior,
it could result in an unusable system.

This patch is intended to force the upgrade of the affected packages
leaving apt no choice but to install additional packages to meet the
dependency requirements rather than uninstalling any packages.